[CHG] dev: nomenclature change ``partner`` -> ``recipient``
This is to prepare the upcoming: ``res.partner`` will not be the
table we are querying, but ``res.partner.backend`` which are actually
recipients (a specific account that can receive money). The word
``recipients`` comes from lokapi nomenclature and it makes sense
to avoid using odoo's backend nomenclature when possible when dealing
with endpoints. Indeed, ``lokapi`` will interface with other
administrative backend, and this API endpoint is to focus on the
a common vocabulary.

This change is also proposed to make the following commits more
readable.

Signed-off-by: Valentin Lab <valentin.lab@kalysto.org>