[FIX] lcc_comchain_base: remove validation of output for ``credit-request``
This method of validation is juged totally useless and cumbersome for
many reasons. The API is still being changed, and datamodels on output
here are difficult to tinker with (does update correctly upon changes).

Their only usage would be to document the API, but this functionality
is not working neither for now.